KVM: selftests: Use real device MSIs in vfio_pci_device_irq_test
Use real device MSIs to deliver interrupts into the guest in
vfio_pci_device_irq_test (when possible). If the device has a driver, we
can use the driver to make the device send a physical MSI, rather than
synthesizing an eventfd notification in software. This allows this test
to exercise IRQ Bypass (e.g. VT-d device-posted interrupts in Intel).

Signed-off-by: David Matlack <dmatlack@google.com>
Change-Id: I23e812b376bd042c155be74a49fe7140b65d2f75