vfio/pci: Support Live Update preserve for VFIO device cdev files
Add the absolute minimum support to plumb a VFIO PCI device file across
a Live Update so that it can be retrieved by userspace after kexec.

Note that this commit only implements the preserve side.  Retrieving the
file after kexec will be added in a subsequent commit.

Create a struct to serialize VFIO PCI data and preserve it using KHO.
Provide physical address of the folio to Live Update Orchestrator (LUO)
in preserve() callback so that LUO can give it back after kexec.
Unpreserve and free the folio in unpreserve() callback.

Store the device's BDF and segment number in the preserved folio. This
data will be used in subsequent commits to correctly reconstruct the
device file when it is retrieved by userspace after kexec.

Require that userspace disables interrupts on the device prior to
freeze() so that the device does not send any interrupts until new
interrupt handlers have been set up by the next kernel.

Reset the device and restore its state just before kexec in the freeze
callback. This ensures the device can be received by the next kernel in
a consistent state. Eventually this should be dropped and the device can
be preserved across in a running state.

Signed-off-by: Vipin Sharma <vipinsh@google.com>
Change-Id: Iac46367cc3a6c470535a710afafd8ee2c2a52ed6