vfio/pci: Allow retrieving preserved device file during Live Update
Retrieve VFIO device in the retrieve() callback of the LUO file handler.
Deserialize the KHO data and search in the VFIO cdev class for device
matching the BDF. Export needed functions from core VFIO module to
others.

Use an anonymous inode when creating the file, since the retrieved
device file is not associated with any particular cdev inode.

Signed-off-by: Vipin Sharma <vipinsh@google.com>
Signed-off-by: David Matlack <dmatlack@google.com>
Change-Id: I836618d9c2b59c71883a283e49731cdee65359fe