PCI: Do not auto-probe incoming Live Update preserved devices
Do not auto-probe devices that were preserved across Live Update from
the previous kernel. Auto-probing could lead to the device being bound
to a different driver than what was bound to the device prior to Live
Update.

Instead, make this userspace's problem and force userspace to set
driver_override to bind the device to a specific driver after Live
Update.

This restriction is relaxed once a driver has called
pci_liveupdate_incoming_finish() on the device.

Signed-off-by: David Matlack <dmatlack@google.com>
Change-Id: I31b91bf816fc490e56d84862e1d29631adc8dcf9